we want an int, so lets tell the compiler to make it not complain
git-svn-id: svn://svn.oetiker.ch/rrdtool/branches/1.2/program@892 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_rpncalc.c
@@ -760,7 +760,7 @@ rpn_calc(rpnp_t *rpnp, rpnstack_t *rpnstack, long data_idx,
            case OP_AVG:
                stackunderflow(0);
                 {
-                   int i=rpnstack -> s[stptr--];
+                   int i=(int)rpnstack -> s[stptr--];
                    double sum=0;
                    int count=0;
                    stackunderflow(i-1);
